functions/get-ai-hint/requirements.txt
appwrite
google-generativeai
